Rework DNS verification call for delegated domains
INF-1540

There was a fundamental mistake in using Dehydrated for calling Letsencrypt
verification in case of delegated domains (in our case, it is beta subdomain
delegated by salemove.com to a separate hosted zone).

Reworked and simplified the Dehydrated-Lexicon junction. Got rid of the
hardcoded hook, instead use a simple search-replace on the default one, when
there is a domain delegation.